Add markdown files that will replace references to diode projects throughout
Previously, there were several links to Diode projects. Diode was a tool that allowed students to create learning goals, but we are deprecating it.

This takes the content of each referenced Diode project and adds it to a markdown file. Any references to the Diode project can be replaced with a link to this markdown file, in a subsequent commit.